ci(integracion):integracion en carpetas respectivas del panel principal
