+++ src/app/private/panelPrincipal/page.tsx
@@ -1,3 +1,4 @@
+// src/app/private/panelPrincipal/page.tsx
 'use client';
 
 import { useEffect, useState } from 'react';
